+++ bindings/python/RobotInterface/src/RobotControl.cpp
@@ -33,6 +33,7 @@ void CreateIRobotControl(pybind11::module& module)
         .value("PositionDirect", IRobotControl::ControlMode::PositionDirect)
         .value("Velocity", IRobotControl::ControlMode::Velocity)
         .value("Torque", IRobotControl::ControlMode::Torque)
+        .value("PWM", IRobotControl::ControlMode::PWM)
         .value("Idle", IRobotControl::ControlMode::Idle)
         .value("Unknown", IRobotControl::ControlMode::Unknown)
         .export_values();