+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project are documented in this file.
 
 ## [Unreleased]
 ### Added
+- Implement `CubicSpline` class (https://github.com/dic-iit/bipedal-locomotion-framework/pull/344)
+- Implement `PWM` control in RobotControl class (https://github.com/dic-iit/bipedal-locomotion-framework/pull/346)
+
+### Changed
+### Fix
+
+## [0.2.0] - 2021-06-15
+### Added
 - Implement IRobotControl python bindings (https://github.com/dic-iit/bipedal-locomotion-framework/pull/200)
 - Implement ISensorBridge python bindings (https://github.com/dic-iit/bipedal-locomotion-framework/pull/203)
 - Implement `LeggedOdometry` class as a part of `FloatingBaseEstimators` library and handle arbitrary contacts in `FloatingBaseEstimator`. (https://github.com/dic-iit/bipedal-locomotion-framework/pull/151)
@@ -62,6 +70,7 @@ All notable changes to this project are documented in this file.
 - Change the parameters handler verbosity (https://github.com/dic-iit/bipedal-locomotion-framework/pull/330)
 - Restore backward compatibility of SwingFootPlanner parameters (https://github.com/dic-iit/bipedal-locomotion-framework/pull/334)
 - Add common Python files to gitignore (https://github.com/dic-iit/bipedal-locomotion-framework/pull/338)
+- Bump manif version to 0.0.4  (https://github.com/dic-iit/bipedal-locomotion-framework/pull/339)
 
 ### Fixed
 - Fix missing implementation of `YarpSensorBridge::getFailedSensorReads()`. (https://github.com/dic-iit/bipedal-locomotion-framework/pull/202)
@@ -116,6 +125,7 @@ All notable changes to this project are documented in this file.
 - Added `mas-imu-test` application to check the output of MAS IMUs (https://github.com/dic-iit/bipedal-locomotion-framework/pull/62)
 - Implement motor currents reading in `YarpSensorBridge`. (https://github.com/dic-iit/bipedal-locomotion-framework/pull/187)
 
-[Unreleased]: https://github.com/dic-iit/bipedal-locomotion-framework/compare/v0.1.1...master
+[Unreleased]: https://github.com/dic-iit/bipedal-locomotion-framework/compare/v0.2.0...master
+[0.2.0]: https://github.com/dic-iit/bipedal-locomotion-framework/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/dic-iit/bipedal-locomotion-framework/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/dic-iit/bipedal-locomotion-framework/releases/tag/v0.1.0